# specs/004-institution-iam-auth-integration/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Clinica Digital IAM API
  version: 0.3.0
  description: Contratos de IAM nativo in-app, registro institucional, sessao opaca, contexto do Shell e validacao RNDS para Organization, Location, Practitioner e PractitionerRole.
servers:
  - url: http://localhost:8080
paths:
...
  /api/admin/tenants:
    post:
      summary: Cria organization e admin (super-user)
      operationId: createTenantAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantAdminRequest'
      responses:
        '201':
          description: Tenant e admin criados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTenantAdminResponse'
        '409':
          description: Duplicidade de nome/CNES/email
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /api/admin/tenants/{tenantId}:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Atualiza dados cadastrais de um tenant existente
      operationId: updateTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSummaryResponse'
        '400':
          description: Requisicao invalida
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Tenant nao encontrado
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
    delete:
      summary: Remove tenant existente
      operationId: deleteTenant
      responses:
        '204':
          description: Tenant removido com sucesso
        '404':
          description: Tenant nao encontrado
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
...
components:
  schemas:
    IdentifierLite:
      type: object
      required: [system, value]
      properties:
        system:
          type: string
          format: uri
        value:
          type: string
        use:
          type: string
    HumanNameLite:
      type: object
      required: [text]
      properties:
        text:
          type: string
        family:
          type: string
        given:
          type: array
          items:
            type: string
    ContactPointLite:
      type: object
      required: [system, value]
      properties:
        system:
          type: string
          enum: [phone, fax, email, pager, url, sms, other]
        value:
          type: string
        use:
          type: string
    AddressLite:
      type: object
      properties:
        line:
          type: array
          items:
            type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    MetaProfile:
      type: object
      required: [profile]
      properties:
        profile:
          type: array
          minItems: 1
          items:
            type: string
            format: uri
...
    PractitionerCreateInput:
      type: object
      required: [displayName, email, cpf, password, identifiers, names, meta]
      properties:
        displayName:
          type: string
        email:
          type: string
          format: email
        cpf:
          type: string
        password:
          type: string
          minLength: 8
        identifiers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IdentifierLite'
        names:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/HumanNameLite'
        telecom:
          type: array
          items:
            $ref: '#/components/schemas/ContactPointLite'
        address:
          type: array
          items:
            $ref: '#/components/schemas/AddressLite'
        meta:
          $ref: '#/components/schemas/MetaProfile'
    OrganizationCreateInput:
      type: object
      required: [name, displayName, cnes, identifiers, meta]
      properties:
        name:
          type: string
        displayName:
          type: string
        cnes:
          type: string
          minLength: 7
          maxLength: 7
        active:
          type: boolean
          default: true
        accountActive:
          type: boolean
          default: true
        identifiers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IdentifierLite'
        telecom:
          type: array
          items:
            $ref: '#/components/schemas/ContactPointLite'
        address:
          type: array
          items:
            $ref: '#/components/schemas/AddressLite'
        meta:
          $ref: '#/components/schemas/MetaProfile'
    CreateTenantAdminRequest:
      type: object
      required: [organization, adminPractitioner]
      properties:
        organization:
          $ref: '#/components/schemas/OrganizationCreateInput'
        adminPractitioner:
          $ref: '#/components/schemas/PractitionerCreateInput'
...
    OrganizationSummary:
      type: object
      required: [id, name, displayName, cnes, identifiers, active, accountActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
        cnes:
          type: string
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/IdentifierLite'
        active:
          type: boolean
        accountActive:
          type: boolean
...
    PractitionerSummary:
      type: object
      required: [id, email, profileType, displayName, accountActive, identifiers, names]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        profileType:
          type: integer
          enum: [0, 10, 20]
        displayName:
          type: string
        accountActive:
          type: boolean
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/IdentifierLite'
        names:
          type: array
          items:
            $ref: '#/components/schemas/HumanNameLite'
...
    CreateTenantAdminResponse:
      type: object
      required: [tenantId, adminPractitionerId, tenant, adminPractitioner]
      properties:
        tenantId:
          type: string
          format: uuid
        adminPractitionerId:
          type: string
          format: uuid
        tenant:
          $ref: '#/components/schemas/OrganizationSummary'
        adminPractitioner:
          $ref: '#/components/schemas/PractitionerSummary'
    TenantSummaryResponse:
      type: object
      required: [id, slug, legalName, status, planTier]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        legalName:
          type: string
        status:
          type: string
        planTier:
          type: string
    UpdateTenantRequest:
      type: object
      required: [slug, legalName, planTier]
      properties:
        slug:
          type: string
        legalName:
          type: string
        planTier:
          type: string
...
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          minItems: 1
          items:
            type: object
            required: [severity, code, details]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
              details:
                type: object
                required: [text]
                properties:
                  text:
                    type: string
              diagnostics:
                type: string